feat: implement display of the current page in the browser URL using query parameters
