+++ rs-react-app/src/main.tsx
@@ -2,13 +2,16 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App.tsx';
+import { BrowserRouter } from 'react-router';
 
 const rootElement = document.getElementById('root');
 
 if (rootElement) {
   createRoot(rootElement).render(
     <StrictMode>
-      <App />
+      <BrowserRouter basename='/REACT2025Q1/'>
+        <App />
+      </BrowserRouter>
     </StrictMode>
   );
 } else {
