+++ rs-react-app/src/pages/About.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const About: React.FC = () => {
+  return <h1>About page</h1>;
+};
+
+export default About;
